Move Stickify folder into a common place - Support folder
