+++ src/Accelerators/NNPA/Conversion/ONNXToZHigh/ONNXLegalityCheck.cpp
@@ -49,7 +49,7 @@ bool isCompatibleWithNNPALevel(std::string inputNNPALevel) {
 
 /// A function to check whether a value's element type is valid for zAIU or not.
 /// zAIU supports only F16, F32 and BFLOAT. Since MLIR does not support BFLOAT,
-/// we check F16 and F32 here only. zAIU only supports rank in range of [0, 4].
+/// we check F16 and F32 here only. zAIU only supports rank in range of (0, 4].
 bool isValidElementTypeAndRank(Value val, bool donotCheckRank) {
   if (val.getType().isa<NoneType>())
     return true;
@@ -65,7 +65,7 @@ bool isValidElementTypeAndRank(Value val, bool donotCheckRank) {
       if (!valueType.hasRank())
         return false;
       int64_t rank = valueType.getRank();
-      return ((rank >= 0) && (rank <= 4));
+      return ((rank > 0) && (rank <= 4));
     }
   }
   return false;
@@ -324,9 +324,9 @@ bool isSuitableForZDNN<ONNXDivOp>(
   // Check NNPA level.
   if (!isCompatibleWithNNPALevel(NNPA_Z16))
     return false;
-  if (!isValidElementTypeAndRank(A))
+  if (!isConstantOfSplatF32Value(A) && !isValidElementTypeAndRank(A))
     return false;
-  if (!isValidElementTypeAndRank(B))
+  if (!isConstantOfSplatF32Value(B) && !isValidElementTypeAndRank(B))
     return false;
   return isConstantOfSplatF32Value(A) || isConstantOfSplatF32Value(B) ||
          dimAnalysis->sameShape(A, B);
