+++ src/Accelerators/NNPA/Support/Stickify/Convert.cpp
@@ -12,9 +12,9 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "src/Accelerators/NNPA/Transform/ZHigh/Stickify/Convert.hpp"
+#include "src/Accelerators/NNPA/Support/Stickify/Convert.hpp"
 #include "src/Accelerators/NNPA/Support/NNPALimit.h"
-#include "src/Accelerators/NNPA/Transform/ZHigh/Stickify/DLF16Conversion.hpp"
+#include "src/Accelerators/NNPA/Support/Stickify/DLF16Conversion.hpp"
 
 /// fp32 -> dlf16 conversion.
 uint64_t fp32_to_dlf16(
